Version 1.3.2.
Major changes:
 + Implementated update_firmware_impl using Task-based class.
 + Removed usage of previous method from sources.
 + Refactored update_firmware_impl and device_container to allow miltiple firmware update processes simultaneously.
 + Added new behaviour to start_firmware_write: upon receiving this command for a device, that already a task pending, this task will be canceled.

Minor changes:
 + Changed get_file_contents signature (size is now long).
 + get_file_contents now allocates memory using new instead of malloc - corresponding changes
 + silenced some log messages in adbk

+++ include/storage.h
@@ -17,6 +17,6 @@ struct IDeviceContainer
 int load_devices(IDeviceContainer* adder);
 int save_devices(DC& container);
 
-char* get_file_contents(const char* name,int *size);
+char* get_file_contents(const char* name,long int *size);
 
 #endif //STORAGE_H
